+++ server/src/resolvers/Subscription.js
@@ -1,0 +1,10 @@
+const Subscription = {
+    update: {
+        subscribe(parent, args, ctx, info) {
+            const { pubsub } = ctx;
+            return pubsub.asyncIterator('update_user');
+        }
+    }
+};
+
+module.exports = Subscription;
