+++ server/src/utils/users.js
@@ -1,0 +1,30 @@
+const users = [
+    {
+        id: '1',
+        name: 'John',
+        age: 30,
+        location: {
+            state: 'New York',
+            city: 'Albany'
+        }
+    },
+    {
+        id: '2',
+        name: 'Mike',
+        age: 35,
+        location: {
+            state: 'North Karelia',
+            city: 'Ylitornio'
+        }
+    },
+    {
+        id: '3',
+        name: 'Jessica',
+        age: 25,
+        location: {
+            state: 'Roraima',
+            city: 'Formosa'
+        }
+    }
+];
+module.exports = users;
